Compute the edges of a cycle.
Following Conway & Ryba, we say that the edges of (abcdef)
are ab,bc,...,fa.

In other words, we identify six-cycles with hexagons, taking the
(ordered) orbit of a single point under the cycle to be the (ordered)
list of points in the hexagon. Then the edges of the cycle are exactly
the edges of the cycle.