Custom action of n-cycles on n-gons; fix bug
We interpret a 6-cycle as a hexagon, taking the (ordered) orbit of one
of the points under the cycle. However, this is different from the
usual action, so if we just do `g(points)` to permute `points`, then we
will get all of the lines, in incorrect order.

This commit also adds in a new test case that checks for this bug.